tests: share the deferral lock harness
